Moved predicate package to pkg/util
