+++ pkg/controller/certificates/informers.go
@@ -25,7 +25,7 @@ import (
 
 	cmlisters "github.com/jetstack/cert-manager/pkg/client/listers/certmanager/v1alpha2"
 	controllerpkg "github.com/jetstack/cert-manager/pkg/controller"
-	"github.com/jetstack/cert-manager/pkg/controller/certificates/internal/predicate"
+	"github.com/jetstack/cert-manager/pkg/util/predicate"
 )
 
 // EnqueueCertificatesForResourceUsingPredicates will return a function
