+++ pkg/controller/certificates/keymanager/BUILD.bazel
@@ -14,9 +14,9 @@ go_library(
         "//pkg/client/listers/certmanager/v1alpha2:go_default_library",
         "//pkg/controller:go_default_library",
         "//pkg/controller/certificates:go_default_library",
-        "//pkg/controller/certificates/internal/predicate:go_default_library",
         "//pkg/logs:go_default_library",
         "//pkg/util/pki:go_default_library",
+        "//pkg/util/predicate:go_default_library",
         "@com_github_go_logr_logr//:go_default_library",
         "@io_k8s_api//core/v1:go_default_library",
         "@io_k8s_apimachinery//pkg/api/errors:go_default_library",
