+++ pkg/controller/certificates/trigger/policies/BUILD.bazel
@@ -12,9 +12,9 @@ go_library(
         "//pkg/apis/certmanager/v1alpha2:go_default_library",
         "//pkg/client/listers/certmanager/v1alpha2:go_default_library",
         "//pkg/controller/certificates:go_default_library",
-        "//pkg/controller/certificates/internal/predicate:go_default_library",
         "//pkg/logs:go_default_library",
         "//pkg/util/pki:go_default_library",
+        "//pkg/util/predicate:go_default_library",
         "@io_k8s_api//core/v1:go_default_library",
         "@io_k8s_apimachinery//pkg/api/errors:go_default_library",
         "@io_k8s_apimachinery//pkg/labels:go_default_library",
