+++ pkg/util/predicate/BUILD.bazel
@@ -8,7 +8,7 @@ go_library(
         "generic.go",
         "predicate.go",
     ],
-    importpath = "github.com/jetstack/cert-manager/pkg/controller/certificates/internal/predicate",
+    importpath = "github.com/jetstack/cert-manager/pkg/util/predicate",
     visibility = ["//pkg/controller/certificates:__subpackages__"],
     deps = [
         "//pkg/apis/certmanager/v1alpha2:go_default_library",
