+++ pkg/util/predicate/generic.go
