+++ pkg/util/predicate/predicate_test.go
